requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp_socks==0.8.4
colorama==0.4.6
curl_cffi==0.6.3
eth_account==0.11.2
loguru==0.7.2
web3==6.19.0
asyncio==3.4.3
fake_useragent==1.5.1
